+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### 🐞 Fixes
 
 - Fixed an issue where aliases would not be inherited for some toolchains.
+- Fixed some token variable interpolation for paths on Windows.
 
 ## 1.32.5